writable5

+++ src/main/java/ca/uwaterloo/cs451/a1/PairsPMI.java
@@ -145,7 +145,7 @@ public class PairsPMI  extends Configured implements Tool {
 
 
   private static final class MyCombiner extends
-      Reducer<PairOfStrings, IntWritable, PairOfStrings, FloatWritable> {
+      Reducer<PairOfStrings, FloatWritable, PairOfStrings, FloatWritable> {
     private static final FloatWritable SUM = new IntWritable();
 
     @Override
